Clean up structure
