+++ .gitignore
@@ -46,3 +46,6 @@ test/rtp_decoder
 test/rtpw
 test/srtp_driver
 test/test_srtp
+
+# Rust things
+test_rs/Cargo.lock
