--- test_rs/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    println!("cargo:rustc-link-search=../build/");
-}
